+++ src/util/p4time.c
@@ -6,11 +6,11 @@
  *
  * \return the number of milliseconds since epoch
  */
-int32_t milli_timer()
+uint64_t milli_timer()
 {
     struct timespec ts;
     clock_gettime(CLOCK_MONOTONIC_RAW, &ts);
 
 
-    return ts.tv_sec * 1000 + (ts.tv_nsec / 1000000UL);
+    return (uint64_t)ts.tv_sec * 1000 + (uint64_t)(ts.tv_nsec / 1000000UL);
 }
